+++ internal/fsx/fsx.go
@@ -1,4 +1,4 @@
-// Package fsx contains file system extension
+// Package fsx contains io/fs extensions.
 package fsx
 
 import (
@@ -8,13 +8,15 @@ import (
 	"syscall"
 )
 
-// Open is a wrapper for os.Open that ensures that we're opening a file.
-func Open(pathname string) (fs.File, error) {
-	return OpenWithFS(filesystem{}, pathname)
+// OpenFile is a wrapper for os.OpenFile that ensures that
+// we're opening a file rather than a directory. If you are
+// opening a directory, this func will return an error.
+func OpenFile(pathname string) (fs.File, error) {
+	return openWithFS(filesystem{}, pathname)
 }
 
-// OpenWithFS is like Open but with explicit file system argument.
-func OpenWithFS(fs fs.FS, pathname string) (fs.File, error) {
+// openWithFS is like Open but with explicit file system argument.
+func openWithFS(fs fs.FS, pathname string) (fs.File, error) {
 	file, err := fs.Open(pathname)
 	if err != nil {
 		return nil, err
